Add Exception

+++ ASPNETCOREAPP/Controllers/HomeController.cs
@@ -5,6 +5,8 @@ using ASPNETCOREAPP.Models;
 using System.IO;
 using Microsoft.AspNetCore.Hosting;
 using System.Linq;
+using System.Threading;
+using ASPNETCOREAPP.Exceptions;
 
 namespace ASPNETCOREAPP.Controllers
 {
@@ -29,22 +31,22 @@ namespace ASPNETCOREAPP.Controllers
         {
             if (ModelState.IsValid)
             {
-                var list = new Listmodel();
+                var listModel = new Listmodel();
 
                 if (model.Photo != null)
                 {
                     var folderPath = Path.Combine(_hostingEnvironment.WebRootPath, "Images");
                     var photoName = Guid.NewGuid() + "_" + model.Photo.FileName;
                     var photoPath = Path.Combine(folderPath, photoName);
-                    list.Photopath = photoName;
+                    listModel.Photopath = photoName;
+                    listModel.Name = model.Name;
 
                     using (var fileStream = new FileStream(photoPath, FileMode.Create))
                     {
                         model.Photo.CopyTo(fileStream); 
                     }
-
-                    list.Name = model.Name;
-                    _dataBase.Listmodels.Add(list);
+                    
+                    _dataBase.Listmodels.Add(listModel);
                     _dataBase.SaveChanges();
                 }
             }
@@ -55,14 +57,21 @@ namespace ASPNETCOREAPP.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(Listmodel model)
         {
-            var entity = await _dataBase.Listmodels.FindAsync(model.ImageId);
+            if (ModelState.IsValid)
+            {
+                var entity = await _dataBase.Listmodels.FindAsync(model.ImageId);
+
+                if (entity is null)
+                throw new EntityNotFoundException();
 
                 var photoPath = Path.Combine(_hostingEnvironment.WebRootPath, "Images", entity.Photopath);
                 FileInfo fileInfo = new FileInfo(photoPath);
                 fileInfo.Delete();
 
-            _dataBase.Listmodels.Remove(entity);
-            await _dataBase.SaveChangesAsync();
+                _dataBase.Listmodels.Remove(entity);
+                await _dataBase.SaveChangesAsync();
+            }
+
             return RedirectToAction(nameof(Index));
         }
     }
